수정

+++ Wafflestagram/app/src/main/java/com/wafflestudio/wafflestagram/ui/main/UserFragment.kt
@@ -83,7 +83,7 @@ class UserFragment: Fragment() {
                     .putExtra("username", userInfo.username)
                     .putExtra("website", userInfo.website)
                     .putExtra("bio", userInfo.bio)
-                    .putExtra("id", userInfo.id)
+                    .putExtra("id", userInfo.id.toInt())
                 startActivity(intent)
             }
